files cleanup
